Extracted ConfigureDeps to separate file.
