+++ CustomerCare/Batch/Runner.cs
@@ -1,5 +1,4 @@
 using System.Threading.Tasks;
-using Raven.Client.Extensions;
 
 namespace CustomerCare.Batch
 {
@@ -7,7 +6,6 @@ namespace CustomerCare.Batch
     using System.Linq;
     using NServiceBus;
     using Raven.Client;
-    using Raven.Client.Document;
     using Raven.Client.Linq;
     using Sales.Contracts;
 
@@ -89,19 +87,6 @@ namespace CustomerCare.Batch
                     });
                 }
             }
-        }
-    }
-
-    public class ConfigureDeps : INeedInitialization
-    {
-        public void Customize(BusConfiguration configuration)
-        {
-            var store = new DocumentStore { Url = "http://localhost:8080", DefaultDatabase = "Customers" };
-            store.Initialize();
-            store.DatabaseCommands.EnsureDatabaseExists("Customers");
-
-            configuration.RegisterComponents(c => c.RegisterSingleton<IDocumentStore>(store));
-            configuration.RegisterComponents(c => c.ConfigureComponent<UpdateCustomerStatusBatch>(DependencyLifecycle.InstancePerCall));
         }
     }
 }
